Added soft deprecation warnings to RxTest. Moved to DeprecationWarner.
